Update ingame-icons.php
Added kibble.  I don't think we need the .kibble_icon_etc at the end because I think this is Unreal's way of saying there may be multiple images which there isn't.  
This will make sure that the additional images work and the file path's are correct.  I could be misunderstanding this and all we would need to do is add it.  And of course, I left the general kibble icon so it doesn't interfere with currently programmed items.  

I then changed the title of SSR to Longneck Rifle.  This will just assure that the title entry changes are taking effect as needed.  

Once this is confirmed, then I will start at the top of the list and work down.